requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
asyncio==3.4.3
pydantic==2.7.0
pydantic_core==2.18.1
marshmallow-sqlalchemy==0.26.1
httpx==0.27.0
aws-secretsmanager-caching==1.1.2
psycopg2-binary==2.9.9
boto3==1.34.36
cryptography==38.0.4
pytz==2022.2.1
asyncpg==0.29.0
greenlet==3.0.3
sqlalchemy-utils==0.41.1
redis==5.0.1
celery==5.3.6